pyproject.toml
[project]
name = "dogbass"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "python-frontmatter>=1.1.0",
]

[project.scripts]
dogbass = "dogbass.cli:main"

[dependency-groups]
dev = [
    "cfn-lint>=1.47.0",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "ruff>=0.15.7",
]
